+++ app/src/main/java/ilya/chistousov/countcalories/domain/model/Food.kt
@@ -6,9 +6,9 @@ data class Food(
     val id: Int = 0,
     val name: String,
     val calories: Int,
-    val proteins: Double,
-    val fats: Double,
-    val carbs: Double,
+    val proteins: Float,
+    val fats: Float,
+    val carbs: Float,
     val meal: Meal,
     val addedDate: LocalDate
 )
